migrate nodesBase to composable
